+++ src/netspresso_trainer/models/heads/detection/__init__.py
@@ -1,3 +1,3 @@
 from .experimental.anchor_free_decoupled_head import anchor_free_decoupled_head
 from .experimental.anchor_decoupled_head import anchor_decoupled_head
-from .experimental.yolov3_head import yolov3_head
+from .experimental.yolo_fastest_head import yolo_fastest_head
